+++ mm2src/coins/sia/src/lib.rs
@@ -1,3 +1,5 @@
+#[macro_use] extern crate serde_json;
+
 pub use ed25519_dalek::{Keypair, PublicKey, SecretKey, Signature};
 
 pub mod blake2b_internal;
@@ -8,3 +10,5 @@ pub mod specifier;
 pub mod spend_policy;
 pub mod types;
 pub mod transaction;
+
+#[cfg(test)] mod tests;
