+++ mm2src/coins/sia/src/tests/mod.rs
@@ -1,4 +1,3 @@
 pub mod encoding;
-pub mod http_client;
 pub mod serde;
 pub mod spend_policy;
