+++ mm2src/coins/sia/src/tests/spend_policy.rs
@@ -1,7 +1,7 @@
-use crate::sia::address::Address;
-use crate::sia::spend_policy::{spend_policy_atomic_swap_success, SpendPolicy, SpendPolicyHelper, UnlockCondition,
+use crate::types::Address;
+use crate::spend_policy::{spend_policy_atomic_swap_success, SpendPolicy, SpendPolicyHelper, UnlockCondition,
                                UnlockKey};
-use crate::sia::PublicKey;
+use crate::PublicKey;
 use rpc::v1::types::H256;
 use std::str::FromStr;
